Refactor file/module structure to separate things
Introduce more separation between parts of the DSL and parts of the
domain model to help tease out a clearer separation between DSL and
template object
